Continue with refactor, move things around.
