--- src/data_types/primitives/mod.rs
@@ -1,11 +1,0 @@
-pub mod bool;
-pub mod dint;
-pub mod int;
-pub mod lreal;
-pub mod real;
-pub mod string16;
-pub mod time;
-pub mod time_struct;
-pub mod udint;
-pub mod uint;
-pub mod word;
